#424 add component that handles redirect to app from notifications
